refactor: floats to double

+++ ws_sensor_combined/src/one_degree_freedom/src/controller.cpp
@@ -26,17 +26,17 @@ class Allocator {
 public:
     // The pwm values are limited by their operating range
     struct ServoAllocatorOutput {
-        float inner_servo_pwm;
-        float outer_servo_pwm;
+        double inner_servo_pwm;
+        double outer_servo_pwm;
     };
 
     struct MotorAllocatorOutput {
-        float upwards_motor_pwm;
-        float downwards_motor_pwm;
+        double upwards_motor_pwm;
+        double downwards_motor_pwm;
     };
 
-    Allocator(float thrust_curve_m, float thrust_curve_b, float g, 
-        float servo_max_tilt_angle_degrees, float motor_max_pwm
+    Allocator(double thrust_curve_m, double thrust_curve_b, double g, 
+        double servo_max_tilt_angle_degrees, double motor_max_pwm
     )   : thrust_curve_m_(thrust_curve_m), thrust_curve_b_(thrust_curve_b), g_(g), 
             servo_max_tilt_angle_degrees_(servo_max_tilt_angle_degrees), motor_max_pwm_(motor_max_pwm)
             {
@@ -54,13 +54,13 @@ public:
             }
 
     ServoAllocatorOutput compute_servo_allocation(
-        float inner_servo_tilt_angle_radians, 
-        float outer_servo_tilt_angle_radians
+        double inner_servo_tilt_angle_radians, 
+        double outer_servo_tilt_angle_radians
     ) {
         ServoAllocatorOutput output;
 
-        float inner_servo_pwm = servo_curve_tilt_radians_to_pwm(inner_servo_tilt_angle_radians);
-        float outer_servo_pwm = servo_curve_tilt_radians_to_pwm(outer_servo_tilt_angle_radians);
+        double inner_servo_pwm = servo_curve_tilt_radians_to_pwm(inner_servo_tilt_angle_radians);
+        double outer_servo_pwm = servo_curve_tilt_radians_to_pwm(outer_servo_tilt_angle_radians);
 
         output.inner_servo_pwm = limit_range_servo_pwm(inner_servo_pwm);
         output.outer_servo_pwm = limit_range_servo_pwm(outer_servo_pwm);
@@ -69,14 +69,14 @@ public:
     }
 
     MotorAllocatorOutput compute_motor_allocation(
-        float delta_motor_pwm, 
-        float average_motor_thrust_newtons
+        double delta_motor_pwm, 
+        double average_motor_thrust_newtons
     ) {
         MotorAllocatorOutput output;
 
-        float average_motor_pwm = motor_thrust_curve_newtons_to_pwm(average_motor_thrust_newtons);
-        float upwards_motor_pwm = average_motor_pwm - delta_motor_pwm / 2.0f;
-        float downwards_motor_pwm = average_motor_pwm + delta_motor_pwm / 2.0f;
+        double average_motor_pwm = motor_thrust_curve_newtons_to_pwm(average_motor_thrust_newtons);
+        double upwards_motor_pwm = average_motor_pwm - delta_motor_pwm / 2.0f;
+        double downwards_motor_pwm = average_motor_pwm + delta_motor_pwm / 2.0f;
 
         output.upwards_motor_pwm = limit_range_motor_pwm(upwards_motor_pwm);
         output.downwards_motor_pwm = limit_range_motor_pwm(downwards_motor_pwm);
@@ -84,38 +84,38 @@ public:
         return output;
     }
 
-    float motor_thrust_curve_pwm_to_newtons(float motor_pwm) {
+    double motor_thrust_curve_pwm_to_newtons(double motor_pwm) {
         return (motor_pwm * thrust_curve_m_ + thrust_curve_b_) / 1000.0f * g_;
     }
 
-    float motor_thrust_curve_newtons_to_pwm(float thrust_newtons) {
+    double motor_thrust_curve_newtons_to_pwm(double thrust_newtons) {
         return ((thrust_newtons * 1000.0f) / g_ - thrust_curve_b_) / thrust_curve_m_;
     }
 
 private: 
-    float thrust_curve_m_;
-    float thrust_curve_b_;
-    float g_;
+    double thrust_curve_m_;
+    double thrust_curve_b_;
+    double g_;
 
-    float servo_max_tilt_angle_degrees_;
-    float motor_max_pwm_;
+    double servo_max_tilt_angle_degrees_;
+    double motor_max_pwm_;
 
-    float limit_range_servo_pwm(float servo_pwm) {
+    double limit_range_servo_pwm(double servo_pwm) {
         servo_pwm = (servo_pwm > 1.0f)? 1.0f : servo_pwm;
         servo_pwm = (servo_pwm < -1.0f)? -1.0f : servo_pwm;
 
         return servo_pwm;
     }
 
-    float limit_range_motor_pwm(float motor_pwm) {
+    double limit_range_motor_pwm(double motor_pwm) {
         motor_pwm = (motor_pwm > motor_max_pwm_)? motor_max_pwm_ : motor_pwm;
         motor_pwm = (motor_pwm < 0.0f)? 0.0f : motor_pwm;
 
         return motor_pwm;
     }
 
-    float servo_curve_tilt_radians_to_pwm(float servo_tilt_angle_radians) {
-        float servo_tilt_angle_degrees = radians_to_degrees(servo_tilt_angle_radians);
+    double servo_curve_tilt_radians_to_pwm(double servo_tilt_angle_radians) {
+        double servo_tilt_angle_degrees = radians_to_degrees(servo_tilt_angle_radians);
         return servo_tilt_angle_degrees / servo_max_tilt_angle_degrees_;
     }
 
@@ -125,12 +125,12 @@ class PIDController
 {
 public:
     // updated asyncronously by the caller
-    std::atomic<float> measurement_;
-    std::atomic<float> measurement_derivative_;
-    std::atomic<float> desired_setpoint_;
-    float computed_output_;
+    std::atomic<double> measurement_;
+    std::atomic<double> measurement_derivative_;
+    std::atomic<double> desired_setpoint_;
+    double computed_output_;
 
-    PIDController(float k_p, float k_d, float k_i, float dt)
+    PIDController(double k_p, double k_d, double k_i, double dt)
         : measurement_{0.0f}, measurement_derivative_{0.0f}, desired_setpoint_{0.0f}, computed_output_(0.0f), 
             k_p_(k_p), k_d_(k_d), k_i_(k_i), integrated_error_(0.0f), dt_(dt)
             {
@@ -155,22 +155,22 @@ public:
         * @brief Compute the control input based on the PID controller formula
         * @return The computed control input
     */
-    float compute() {
+    double compute() {
         // Update the integrated error
         integrated_error_ = integrated_error_ + (desired_setpoint_ - measurement_) * dt_; 
 
         // Compute control input
-        float dot_product = measurement_ * k_p_ + measurement_derivative_ * k_d_;
+        double dot_product = measurement_ * k_p_ + measurement_derivative_ * k_d_;
         computed_output_ = -dot_product + integrated_error_ * k_i_;
         return computed_output_;
     }
 
 private:
-    const float k_p_;
-    const float k_d_;
-    const float k_i_;
-    float integrated_error_;
-    const float dt_;
+    const double k_p_;
+    const double k_d_;
+    const double k_i_;
+    double integrated_error_;
+    const double dt_;
 };
 
 /**
@@ -242,16 +242,16 @@ public:
         }
     )}
     {
-        this->declare_parameter<float>(CONTROLLER_FREQUENCY_HERTZ_PARAM);
-        float controllers_freq = this->get_parameter(CONTROLLER_FREQUENCY_HERTZ_PARAM).as_double();
+        this->declare_parameter<double>(CONTROLLER_FREQUENCY_HERTZ_PARAM);
+        double controllers_freq = this->get_parameter(CONTROLLER_FREQUENCY_HERTZ_PARAM).as_double();
         if (controllers_freq <= 0.0f || controllers_freq == NAN) {
             RCLCPP_ERROR(this->get_logger(), "Could not read controller frequency correctly.");
             throw std::runtime_error("Controller frequency invalid");
         }
         RCLCPP_INFO(this->get_logger(), "Controllers freq: %f", controllers_freq);
-        float controllers_dt = 1.0 / controllers_freq;
+        double controllers_dt = 1.0 / controllers_freq;
 
-        this->declare_parameter<float>(CONTROLLER_DEFAULT_MOTOR_PWM);
+        this->declare_parameter<double>(CONTROLLER_DEFAULT_MOTOR_PWM);
         default_motor_pwm_ = this->get_parameter(CONTROLLER_DEFAULT_MOTOR_PWM).as_double();
         if (default_motor_pwm_ < 0.0f || default_motor_pwm_ > 1.0f || default_motor_pwm_ == NAN) {
             RCLCPP_ERROR(this->get_logger(), "Could not read motor thrust correctly.");
@@ -260,16 +260,16 @@ public:
         RCLCPP_INFO(this->get_logger(), "Default motor thrust pwm: %f", default_motor_pwm_);
 
 
-        this->declare_parameter<float>(GRAVITATIONAL_ACCELERATION);
-        this->declare_parameter<float>(CONTROLLER_THRUST_CURVE_M_PARAM);
-        this->declare_parameter<float>(CONTROLLER_THRUST_CURVE_B_PARAM);
-        this->declare_parameter<float>(CONTROLLER_SERVO_MAX_TILT_ANGLE_PARAM);
-        this->declare_parameter<float>(CONTROLLER_MOTOR_MAX_PWM_PARAM);
-        float thrust_curve_m = this->get_parameter(CONTROLLER_THRUST_CURVE_M_PARAM).as_double();
-        float thrust_curve_b = this->get_parameter(CONTROLLER_THRUST_CURVE_B_PARAM).as_double();
-        float g               = this->get_parameter(GRAVITATIONAL_ACCELERATION).as_double();
-        float servo_max_tilt_angle_degrees = this->get_parameter(CONTROLLER_SERVO_MAX_TILT_ANGLE_PARAM).as_double();
-        float motor_max_pwm   = this->get_parameter(CONTROLLER_MOTOR_MAX_PWM_PARAM).as_double();
+        this->declare_parameter<double>(GRAVITATIONAL_ACCELERATION);
+        this->declare_parameter<double>(CONTROLLER_THRUST_CURVE_M_PARAM);
+        this->declare_parameter<double>(CONTROLLER_THRUST_CURVE_B_PARAM);
+        this->declare_parameter<double>(CONTROLLER_SERVO_MAX_TILT_ANGLE_PARAM);
+        this->declare_parameter<double>(CONTROLLER_MOTOR_MAX_PWM_PARAM);
+        double thrust_curve_m = this->get_parameter(CONTROLLER_THRUST_CURVE_M_PARAM).as_double();
+        double thrust_curve_b = this->get_parameter(CONTROLLER_THRUST_CURVE_B_PARAM).as_double();
+        double g               = this->get_parameter(GRAVITATIONAL_ACCELERATION).as_double();
+        double servo_max_tilt_angle_degrees = this->get_parameter(CONTROLLER_SERVO_MAX_TILT_ANGLE_PARAM).as_double();
+        double motor_max_pwm   = this->get_parameter(CONTROLLER_MOTOR_MAX_PWM_PARAM).as_double();
 
         allocator_ = std::make_unique<Allocator>(
             thrust_curve_m, thrust_curve_b, g, 
@@ -279,13 +279,13 @@ public:
         this->declare_parameter<bool>(CONTROLLER_ROLL_ACTIVE_PARAM);
         if (this->get_parameter(CONTROLLER_ROLL_ACTIVE_PARAM).as_bool()) {
             RCLCPP_INFO(this->get_logger(), "Roll is active");
-            this->declare_parameter<float>(CONTROLLER_ROLL_K_P_PARAM);
-            this->declare_parameter<float>(CONTROLLER_ROLL_K_D_PARAM);
-            this->declare_parameter<float>(CONTROLLER_ROLL_K_I_PARAM);
+            this->declare_parameter<double>(CONTROLLER_ROLL_K_P_PARAM);
+            this->declare_parameter<double>(CONTROLLER_ROLL_K_D_PARAM);
+            this->declare_parameter<double>(CONTROLLER_ROLL_K_I_PARAM);
 
-            float k_p = this->get_parameter(CONTROLLER_ROLL_K_P_PARAM).as_double();
-            float k_d = this->get_parameter(CONTROLLER_ROLL_K_D_PARAM).as_double();
-            float k_i = this->get_parameter(CONTROLLER_ROLL_K_I_PARAM).as_double();
+            double k_p = this->get_parameter(CONTROLLER_ROLL_K_P_PARAM).as_double();
+            double k_d = this->get_parameter(CONTROLLER_ROLL_K_D_PARAM).as_double();
+            double k_i = this->get_parameter(CONTROLLER_ROLL_K_I_PARAM).as_double();
 
             roll_controller_.emplace(k_p, k_d, k_i, controllers_dt);
         } else {
@@ -297,13 +297,13 @@ public:
         this->declare_parameter<bool>(CONTROLLER_PITCH_ACTIVE_PARAM);
         if (this->get_parameter(CONTROLLER_PITCH_ACTIVE_PARAM).as_bool()) {
             RCLCPP_INFO(this->get_logger(), "Pitch is active");
-            this->declare_parameter<float>(CONTROLLER_PITCH_K_P_PARAM);
-            this->declare_parameter<float>(CONTROLLER_PITCH_K_D_PARAM);
-            this->declare_parameter<float>(CONTROLLER_PITCH_K_I_PARAM);
+            this->declare_parameter<double>(CONTROLLER_PITCH_K_P_PARAM);
+            this->declare_parameter<double>(CONTROLLER_PITCH_K_D_PARAM);
+            this->declare_parameter<double>(CONTROLLER_PITCH_K_I_PARAM);
 
-            float k_p = this->get_parameter(CONTROLLER_PITCH_K_P_PARAM).as_double();
-            float k_d = this->get_parameter(CONTROLLER_PITCH_K_D_PARAM).as_double();
-            float k_i = this->get_parameter(CONTROLLER_PITCH_K_I_PARAM).as_double();
+            double k_p = this->get_parameter(CONTROLLER_PITCH_K_P_PARAM).as_double();
+            double k_d = this->get_parameter(CONTROLLER_PITCH_K_D_PARAM).as_double();
+            double k_i = this->get_parameter(CONTROLLER_PITCH_K_I_PARAM).as_double();
 
             pitch_controller_.emplace(k_p, k_d, k_i, controllers_dt);
         } else {
@@ -315,13 +315,13 @@ public:
         this->declare_parameter<bool>(CONTROLLER_YAW_ACTIVE_PARAM);
         if (this->get_parameter(CONTROLLER_YAW_ACTIVE_PARAM).as_bool()) {
             RCLCPP_INFO(this->get_logger(), "Yaw is active");
-            this->declare_parameter<float>(CONTROLLER_YAW_K_P_PARAM);
-            this->declare_parameter<float>(CONTROLLER_YAW_K_D_PARAM);
-            this->declare_parameter<float>(CONTROLLER_YAW_K_I_PARAM);
+            this->declare_parameter<double>(CONTROLLER_YAW_K_P_PARAM);
+            this->declare_parameter<double>(CONTROLLER_YAW_K_D_PARAM);
+            this->declare_parameter<double>(CONTROLLER_YAW_K_I_PARAM);
 
-            float k_p = this->get_parameter(CONTROLLER_YAW_K_P_PARAM).as_double();
-            float k_d = this->get_parameter(CONTROLLER_YAW_K_D_PARAM).as_double();
-            float k_i = this->get_parameter(CONTROLLER_YAW_K_I_PARAM).as_double();
+            double k_p = this->get_parameter(CONTROLLER_YAW_K_P_PARAM).as_double();
+            double k_d = this->get_parameter(CONTROLLER_YAW_K_D_PARAM).as_double();
+            double k_i = this->get_parameter(CONTROLLER_YAW_K_I_PARAM).as_double();
 
             yaw_controller_.emplace(k_p, k_d, k_i, controllers_dt);
         } else {
@@ -331,7 +331,7 @@ public:
 
 
         controller_timer_ = this->create_wall_timer(
-            std::chrono::duration<float>(controllers_dt), 
+            std::chrono::duration<double>(controllers_dt), 
             std::bind(&BaselinePIDController::controller_callback, this)
         );
 	}
@@ -360,25 +360,25 @@ private:
 
     std::unique_ptr<Allocator> allocator_;
 
-    float default_motor_pwm_;
+    double default_motor_pwm_;
 
 	//!< Auxiliary functions
     void controller_callback();
     void publish_controller_debug();
     void publish_allocator_debug(
-        const float inner_servo_tilt_angle_radians,
-        const float outer_servo_tilt_angle_radians,
-        const float inner_servo_pwm,
-        const float outer_servo_pwm,
+        const double inner_servo_tilt_angle_radians,
+        const double outer_servo_tilt_angle_radians,
+        const double inner_servo_pwm,
+        const double outer_servo_pwm,
 
-        const float delta_motor_pwm,
-        const float average_motor_thrust_newtons,
-        const float upwards_motor_pwm,
-        const float downwards_motor_pwm
+        const double delta_motor_pwm,
+        const double average_motor_thrust_newtons,
+        const double upwards_motor_pwm,
+        const double downwards_motor_pwm
     );
 
-    void publish_servo_pwm(const float inner_servo_pwm, const float outer_servo_pwm);
-    void publish_motor_pwm(const float upwards_motor_pwm, const float downwards_motor_pwm);
+    void publish_servo_pwm(const double inner_servo_pwm, const double outer_servo_pwm);
+    void publish_motor_pwm(const double upwards_motor_pwm, const double downwards_motor_pwm);
 
     std::atomic<uint8_t> flight_mode_;
     rclcpp::Subscription<FlightMode>::SharedPtr flight_mode_get_subscriber_;
@@ -392,10 +392,10 @@ void BaselinePIDController::controller_callback()
 {
     // only run controller when in mission
     if (flight_mode_.load() == FlightMode::IN_MISSION) {
-        float inner_servo_tilt_angle_radians = (roll_controller_)  ? roll_controller_->compute() : 0.0f;
-        float outer_servo_tilt_angle_radians = (pitch_controller_) ? pitch_controller_->compute() : 0.0f;
-        float delta_motor_pwm = (yaw_controller_) ? yaw_controller_->compute() : 0.0f;
-        float average_motor_thrust_newtons = allocator_->motor_thrust_curve_pwm_to_newtons(default_motor_pwm_);
+        double inner_servo_tilt_angle_radians = (roll_controller_)  ? roll_controller_->compute() : 0.0f;
+        double outer_servo_tilt_angle_radians = (pitch_controller_) ? pitch_controller_->compute() : 0.0f;
+        double delta_motor_pwm = (yaw_controller_) ? yaw_controller_->compute() : 0.0f;
+        double average_motor_thrust_newtons = allocator_->motor_thrust_curve_pwm_to_newtons(default_motor_pwm_);
 
         Allocator::ServoAllocatorOutput servo_output = allocator_->compute_servo_allocation(
             inner_servo_tilt_angle_radians,
@@ -464,15 +464,15 @@ void BaselinePIDController::publish_controller_debug() {
 }
 
 void BaselinePIDController::publish_allocator_debug(
-    const float inner_servo_tilt_angle_radians,
-    const float outer_servo_tilt_angle_radians,
-    const float inner_servo_pwm,
-    const float outer_servo_pwm,
+    const double inner_servo_tilt_angle_radians,
+    const double outer_servo_tilt_angle_radians,
+    const double inner_servo_pwm,
+    const double outer_servo_pwm,
 
-    const float delta_motor_pwm,
-    const float average_motor_thrust_newtons,
-    const float upwards_motor_pwm,
-    const float downwards_motor_pwm
+    const double delta_motor_pwm,
+    const double average_motor_thrust_newtons,
+    const double upwards_motor_pwm,
+    const double downwards_motor_pwm
 ) {
     AllocatorDebug msg{};
     msg.stamp = this->get_clock()->now();
@@ -490,7 +490,7 @@ void BaselinePIDController::publish_allocator_debug(
     allocator_debug_publisher_->publish(msg);
 }
 
-void BaselinePIDController::publish_servo_pwm(const float inner_servo_pwm, const float outer_servo_pwm)
+void BaselinePIDController::publish_servo_pwm(const double inner_servo_pwm, const double outer_servo_pwm)
 {
     ActuatorServos msg{};
 	msg.timestamp = this->get_clock()->now().nanoseconds() / 1000;
@@ -503,7 +503,7 @@ void BaselinePIDController::publish_servo_pwm(const float inner_servo_pwm, const
  * @brief Publish the actuator motors.
  *        For this example, we are generating sinusoidal values for the actuator positions.
  */
-void BaselinePIDController::publish_motor_pwm(const float upwards_motor_pwm, const float downwards_motor_pwm)
+void BaselinePIDController::publish_motor_pwm(const double upwards_motor_pwm, const double downwards_motor_pwm)
 {
 	ActuatorMotors msg{};
 	msg.timestamp = this->get_clock()->now().nanoseconds() / 1000;
